# Conflicts:
#	docs/docs/reference/api.mdx
#	internal/httpclient/models/complete_self_service_login_flow_with_password_method.go
#	internal/httpclient/models/complete_self_service_recovery_flow_with_link_method.go
#	internal/httpclient/models/complete_self_service_settings_flow_with_password_method.go
#	internal/httpclient/models/complete_self_service_verification_flow_with_link_method.go
#	internal/httpclient/models/create_identity.go
#	internal/httpclient/models/create_recovery_link.go
#	internal/httpclient/models/form_field.go
#	internal/httpclient/models/generic_error_payload.go
#	internal/httpclient/models/id.go
#	internal/httpclient/models/identity.go
#	internal/httpclient/models/message.go
#	internal/httpclient/models/messages.go
#	internal/httpclient/models/recovery_address.go
#	internal/httpclient/models/recovery_address_type.go
#	internal/httpclient/models/registration_flow.go
#	internal/httpclient/models/registration_flow_method_config.go
#	internal/httpclient/models/revoke_session.go
#	internal/httpclient/models/settings_via_api_response.go
#	internal/httpclient/models/state.go
#	internal/httpclient/models/traits.go
#	internal/httpclient/models/update_identity.go
#	internal/httpclient/models/verifiable_address.go
#	internal/httpclient/models/verifiable_address_status.go
#	internal/httpclient/models/verifiable_address_type.go
#	internal/httpclient/models/verification_flow.go
#	internal/httpclient/models/version.go

+++ identity/pool.go
@@ -147,7 +147,7 @@ func TestPool(ctx context.Context, conf *config.Config, p interface {
 
 			count, err := p.CountIdentities(ctx)
 			require.NoError(t, err)
-			assert.EqualValues(t, 1, count)
+			assert.EqualValues(t, int64(1), count)
 		})
 
 		t.Run("case=create with default values", func(t *testing.T) {
